Obhajoba

+++ www/vitd04/sp/apps/nextjs/src/app/(app)/(authenticated)/settings/(detail)/restaurants/[slug]/edit-menu/MenuEditor.tsx
@@ -71,10 +71,12 @@ export function MenuEditor({ restaurant }: Props) {
 
   const router = useRouter();
   const handleDeleteItem = (menuItem: number) => {
-    deleteItem({
-      id: menuItem,
-    });
-    router.refresh();
+    if (confirm("Opravdu chcete smazat tuto položku?")) {
+      deleteItem({
+        id: menuItem,
+      });
+      router.refresh();
+    }
   };
 
   // Delete section
